+++ lambda/etl/tests/conftest.py
@@ -28,41 +28,6 @@ def create_test_bucket(local_stack_container: LocalStackContainer) -> None:
     client.create_bucket(Bucket=bucket_name, CreateBucketConfiguration={"LocationConstraint": local_stack_container.region_name})
 
 
-@pytest.fixture
-def valid_assets_page() -> str:
-    """テスト用の正常なHTMLファイルを読み込む"""
-    with open("tests/fixtures/html/valid_assets_page.html") as f:
-        assets_page = f.read()
-    return assets_page
-
-
-@pytest.fixture
-def invalid_assets_page() -> str:
-    """テスト用の不正なHTMLファイルを読み込む"""
-    with open("tests/fixtures/html/invalid_assets_page.html") as f:
-        assets_page = f.read()
-    return assets_page
-
-
-@pytest.fixture
-def valid_assets_info():
-    """テスト用のDcpAssetsInfoオブジェクトを生成する"""
-    from src.domain.value_object import DcpAssetsInfo, DcpProductAssets, DcpTotalAssets
-    return DcpAssetsInfo(
-        total=DcpTotalAssets(
-            cumulative_contributions="900,000円", total_gains_or_losses="300,000円", total_asset_valuation="1,200,000円"
-        ),
-        products={
-            "プロダクト_1": DcpProductAssets(
-                cumulative_acquisition_costs="100,000円", gains_or_losses="11,111円", asset_valuation="111,111円"
-            ),
-            "プロダクト_2": DcpProductAssets(
-                cumulative_acquisition_costs="200,000円", gains_or_losses="22,222円", asset_valuation="222,222円"
-            ),
-        },
-    )
-
-
 @pytest.fixture(scope="function")
 def put_login_parameter(local_stack_container: LocalStackContainer) -> None:
     parameter_name="/test/parameter"
